use style
